+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 
 - Invoking parameter functions of `alternation` twice on checking for its result - now they invokes only once.
+- Verbose type `Maybe`.
 
 ## [0.4.0] - 2020-08-12